refactor(#8/user): 유저 관련 exception을 application -> domain으로 변경
