+++ src/main/java/org/sopt/comment/application/command/CommentCommandService.java
@@ -3,7 +3,7 @@ package org.sopt.comment.application.command;
 import static org.sopt.comment.domain.exception.CommentErrorCode.COMMENT_NOT_FOUND;
 import static org.sopt.comment.domain.exception.CommentErrorCode.COMMENT_UPDATE_UNAUTHORIZED;
 import static org.sopt.post.domain.exception.PostErrorCode.POST_NOT_FOUND;
-import static org.sopt.user.application.exception.UserErrorCode.USER_NOT_FOUND;
+import static org.sopt.user.domain.exception.UserErrorCode.USER_NOT_FOUND;
 
 import lombok.RequiredArgsConstructor;
 import org.sopt.comment.application.dto.request.CreateCommentServiceRequest;
